store: Pass valid integer to assertEquals() test when using $delta
This worked by accident in PHPUnit 3.x, but is unsupported in
PHPUnit 4.x where it will just result in an assertion failure due
to the strings being different. (T99982, I70ad8151)

+++ repo/tests/phpunit/includes/store/sql/SqlChangeStoreTest.php
@@ -109,8 +109,9 @@ class SqlChangeStoreTest extends \MediaWikiTestCase {
 		$this->assertTrue( is_numeric( $row['change_id'] ) );
 
 		$this->assertEquals(
-			wfTimestamp( TS_UNIX, $expected['change_time'] ),
-			wfTimestamp( TS_UNIX, $row['change_time'] ),
+			// wfTimestamp returns string, assertEquals/$delta requires int
+			(int) wfTimestamp( TS_UNIX, $expected['change_time'] ),
+			(int) wfTimestamp( TS_UNIX, $row['change_time'] ),
 			'Change time',
 			60 * 60 // 1 hour
 		);
